pyproject.toml
[tool.poetry]
name = "celpoints"
version = "0.1.0"
description = ""
authors = ["celeste <parafactual@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.1.0"
dataset = "^1.5.2"
python-dotenv = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
